Mongodb issues fixed + Spring API REST functional with MongoDB

+++ src/main/java/com/dicegame/mongodb/dao/IDiceRollDAO.java
@@ -3,6 +3,6 @@ package com.dicegame.mongodb.dao;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import com.dicegame.mongodb.dto.DiceRoll;
 
-public interface IDiceRollDAO extends MongoRepository<DiceRoll, Long> {
+public interface IDiceRollDAO extends MongoRepository<DiceRoll, Integer> {
 
 }
