+++ src/main/java/com/dicegame/mongodb/dao/IPlayerDAO.java
@@ -3,6 +3,6 @@ package com.dicegame.mongodb.dao;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import com.dicegame.mongodb.dto.Player;
 
-public interface IPlayerDAO extends MongoRepository<Player, Long> {
+public interface IPlayerDAO extends MongoRepository<Player, Integer> {
 
 }
